+++ Flow/src/utopia/flow/time/Year.scala
@@ -1,12 +1,13 @@
 package utopia.flow.time
 
-import utopia.flow.operator.{Reversible, Steppable}
 import utopia.flow.operator.combine.{Combinable, Subtractable}
 import utopia.flow.operator.enumeration.Extreme
 import utopia.flow.operator.ordering.SelfComparable
 import utopia.flow.operator.sign.Sign
+import utopia.flow.operator.{Reversible, Steppable}
 
 import java.time
+import java.time.LocalDate
 import scala.language.implicitConversions
 
 object Year
@@ -60,6 +61,19 @@ case class Year(value: Int)
 	def isLeap = java.time.Year.isLeap(value)
 	
 	/**
+	  * @return The first day of this year (i.e. 1st of January)
+	  */
+	def firstDay = LocalDate.of(value, 1, 1)
+	/**
+	  * @return The last day of this year (i.e. 31st of December)
+	  */
+	def lastDay = LocalDate.of(value, 12, 31)
+	/**
+	  * @return Dates of this year
+	  */
+	def dates = DateRange.inclusive(firstDay, lastDay)
+	
+	/**
 	 * @return A Java model of this year
 	 */
 	def toJava = time.Year.of(value)
@@ -77,4 +91,52 @@ case class Year(value: Int)
 	
 	override def next(direction: Sign): Year = Year(value + direction.modifier)
 	override def is(extreme: Extreme): Boolean = false
+	
+	
+	// OTHER    -----------------------------
+	
+	/**
+	  * @param month Targeted month
+	  * @return That month at this year
+	  */
+	def apply(month: Month) = YearMonth(this, month)
+	/**
+	  * @param monthOfYear Value of the targeted month [1,12]
+	  * @return That month at this year
+	  */
+	@throws[IndexOutOfBoundsException]("If 'monthOfYear' is out of range")
+	def apply(monthOfYear: Int): YearMonth = apply(Month(monthOfYear))
+	/**
+	  * @param monthDay Targeted month & day of month
+	  * @return That day at this year
+	  */
+	def apply(monthDay: MonthDay) = LocalDate.of(value, monthDay.month.value, monthDay.day)
+	/**
+	  * @param range A range of dates
+	  * @return The portion of that range that overlaps with this year. 0-2 different ranges.
+	  */
+	def apply(range: YearlyDateRange): IndexedSeq[DateRange] = range.at(this)
+	
+	/**
+	  * Adds month information to this year
+	  * @param month Targeted month
+	  * @return A monthYear based on this year and specified month
+	  */
+	def +(month: Month) = apply(month)
+	/**
+	  * Adds month information (same as this + month)
+	  * @param month targeted month
+	  * @return targeted month on this year
+	  */
+	def /(month: Month) = apply(month)
+	/**
+	  * @param monthDay A month day
+	  * @return That month day during this year
+	  */
+	def /(monthDay: MonthDay) = apply(monthDay)
+	/**
+	  * @param range A range of dates
+	  * @return The portion of that range that overlaps with this year. 0-2 different ranges.
+	  */
+	def /(range: YearlyDateRange) = apply(range)
 }